Improve address type naming
